+++ CHANGES.rst
@@ -63,6 +63,8 @@ Improvements
 - Support attaching files when emailing event persons (:pr:`7369`, thanks :user:`jbtwist`)
 - Let registration managers view logs for specific registrations (:pr:`7600`, thanks
   :user:`vtran99`)
+- Add pagination to public participant list and use tabs instead of an accordion in case
+  of multiple registration forms (:issue:`6424`, :pr:`7472`)
 
 Bugfixes
 ^^^^^^^^